requirements.txt
asgiref==3.8.1
astroid==3.2.3
certifi==2024.7.4
charset-normalizer==3.3.2
dill==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
platformdirs==4.2.2
PyJWT==2.8.0
pylint==3.2.5
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.0
urllib3==2.2.2
